requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0

# Tracking
boxmot>=10.0.0
filterpy>=1.4.5

# Data processing
scikit-learn>=1.3.0
pyyaml>=6.0.0

# Optional but recommended
tqdm>=4.66.0
matplotlib>=3.7.0
